mcp-servers/Speak-Identification/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sse-starlette>=2.0.0
mcp>=0.1.0
fastmcp==2.9.0
requests==2.32.4
google-generativeai==0.8.5
pytz==2022.1